% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baum_welch.R
\name{EPI.genefinder}
\alias{EPI.genefinder}
\title{Forward-backward algorithm}
\usage{
EPI.genefinder(clzz, emissionPars, initP = NULL, transitionProbs,
  seqlens = nrow(clzz), updateEmission = NULL,
  updateTransition = NULL, maxiter = 10, tol = 1e-04, nthreads = 1)
}
\arguments{
\item{clzz}{binary matrix of histone marks}

\item{emissionPars}{matrix with emission probabilities for each datapoint and each state.}

\item{initP}{matrix of initial probabilities: each column corresponds to a sequence}

\item{transitionProbs}{transition matrix (rows are previous state, columns are next state)
Columns are datapoints and rows are states.}

\item{seqlens}{named list containing the length of chromosomes/contigs.}

\item{updateEmission}{update emission vector with length is same as the number of states, this vector contains boolean values (TRUE or FALSE)}

\item{updateTransition}{update emission matrix with same dimensions as emission probability matrix containing boolean value (TRUE or FALSE)}

\item{maxiter}{number of iterations.}

\item{nthreads}{number of threads used. Sequences of observations are
processed independently by different threads (if \code{length(seqlens) > 1}).}
}
\value{
a list with the following arguments:
   \item{post}{posterior probability of being in a certain state for a certain datapoint.
    Same matrix used as input argument.}
   \item{tot_llik}{total log-likelihood of the data given the hmm model.}
   \item{new_trans}{update for the transition probabilities (it is already normalized).}
   \item{epar}{estimated parameters}
}
\description{
Forward-backward algorithm using the scaling technique.
That's more stable (and maybe even faster) than the method with the logarithm.
Warning: this function overwrites the lliks matrix.
}
